+++ Atlas.MatchingAlgorithm.Test.Integration/DependencyInjection/ServiceConfiguration.cs
@@ -2,7 +2,7 @@ using Atlas.Common.ApplicationInsights;
 using Atlas.Common.Caching;
 using Atlas.Common.FeatureManagement;
 using Atlas.Common.Notifications;
-using Atlas.Common.ServiceBus.BatchReceiving;
+using Atlas.Common.ServiceBus;
 using Atlas.Common.Test.SharedTestHelpers;
 using Atlas.DonorImport.ExternalInterface;
 using Atlas.DonorImport.ExternalInterface.Models;
